@@ -1,9 +1,14 @@
 package guru.springframework.domain;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.util.HashSet;
 import java.util.Set;
 
+@Data
 @Entity
+@EqualsAndHashCode(exclude = {"recipes"})
 public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -14,19 +19,4 @@ public class Category {
     @ManyToMany(mappedBy = "categories")
     private Set<Recipe> recipes;
 
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
 }
